Extract `toc/poin_ops.rs`
